card11/requirements.txt
google-generativeai
google-adk
litellm

fastapi
uvicorn[standard]

httpx

streamlit

pydantic

python-dotenv
